+++ src/app/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod add;
 pub(crate) mod config_path;
+pub(crate) mod disable;
 pub(crate) mod edit;
 pub(crate) mod enable;
 pub(crate) mod init;