+++ src/core/disable.rs
@@ -1,4 +1,3 @@
-use super::add::add_alias_str;
 use super::list::get_aliases_from_single_group;
 use super::{Failure, Outcome};
 
@@ -8,7 +7,7 @@ use crate::app::shell::ShellType;
 
 use log::error;
 
-pub fn enable_alias(config: &mut Config, name: &str) -> Result<Outcome, Failure> {
+pub fn disable_alias(config: &mut Config, name: &str) -> Result<Outcome, Failure> {
     if !config.aliases.contains_key(name) {
         error!("Alias {} does not exist.", name);
         return Err(Failure::AliasDoesNotExist);
@@ -16,24 +15,24 @@ pub fn enable_alias(config: &mut Config, name: &str) -> Result<Outcome, Failure>
 
     let alias = config.aliases.get_mut(name).unwrap();
 
-    if alias.enabled {
+    if !alias.enabled {
         return Ok(Outcome::NoChanges);
     }
 
-    alias.enabled = true;
+    alias.enabled = false;
 
     // Checks if the group the alias is in is disabled
-    // If it is, then the alias will not be added to the shell
+    // If it is, then the alias will not be removed from the shell
     if let Some(group) = &alias.group
         && !config.groups[group]
     {
         return Ok(Outcome::ConfigChanged);
     }
 
-    Ok(Outcome::Command(add_alias_str(name, alias)))
+    Ok(Outcome::Command(format!("unalias '{}'", name)))
 }
 
-pub fn enable_group(
+pub fn disable_group(
     config: &mut Config,
     name: &str,
     shell: &ShellType,
@@ -43,12 +42,14 @@ pub fn enable_group(
         return Err(Failure::GroupDoesNotExist);
     }
 
-    if config.groups[name] {
+    // If the group is already disabled, do nothing
+    if !config.groups[name] {
         return Ok(Outcome::NoChanges);
     }
 
-    *config.groups.get_mut(name).unwrap() = true;
+    *config.groups.get_mut(name).unwrap() = false;
 
+    // Get all aliases in the group that are enabled and remove them from the shell
     let mut aliases_in_group = get_aliases_from_single_group(config, Some(name), shell)?;
     aliases_in_group.retain(|alias_name| config.aliases[alias_name].enabled);
 
@@ -58,8 +59,7 @@ pub fn enable_group(
 
     let mut command = String::new();
     for alias_name in aliases_in_group {
-        let alias = &config.aliases[&alias_name];
-        command.push_str(&add_alias_str(&alias_name, alias));
+        command.push_str(&format!("unalias '{}'\n", alias_name));
         command.push('\n');
     }
 
@@ -77,102 +77,101 @@ mod test {
         let mut config = Config::new();
         config.groups.insert("enabled_group".into(), true);
         config.groups.insert("disabled_group".into(), false);
-        config.groups.insert("empty_group".into(), false);
+        config.groups.insert("empty_group".into(), true);
 
         config.aliases.insert(
             "alias1".into(),
-            Alias::new("cmd".into(), Some("enabled_group".into()), false, false),
+            Alias::new("cmd".into(), Some("enabled_group".into()), true, false),
         );
         config.aliases.insert(
             "alias2".into(),
-            Alias::new("cmd".into(), Some("disabled_group".into()), false, false),
+            Alias::new("cmd".into(), Some("disabled_group".into()), true, false),
         );
 
         config
     }
 
     #[test]
-    fn enable_existing_alias() {
+    fn disable_existing_alias() {
         let mut config = sample_config();
-        let result = enable_alias(&mut config, "alias1");
+        let result = disable_alias(&mut config, "alias1");
         assert!(result.is_ok());
-        assert!(config.aliases["alias1"].enabled);
+        assert!(!config.aliases["alias1"].enabled);
         assert_matches!(result.unwrap(), Outcome::Command(_));
     }
 
     #[test]
-    fn enable_enabled_alias() {
+    fn disable_disabled_alias() {
         let mut config = sample_config();
-        let _ = enable_alias(&mut config, "alias1");
-        assert!(config.aliases["alias1"].enabled);
+        let _ = disable_alias(&mut config, "alias1");
+        assert!(!config.aliases["alias1"].enabled);
 
-        let result = enable_alias(&mut config, "alias1");
+        let result = disable_alias(&mut config, "alias1");
         assert!(result.is_ok());
-        assert!(config.aliases["alias1"].enabled);
-        assert_eq!(result.unwrap(), Outcome::NoChanges);
+        assert!(!config.aliases["alias1"].enabled);
+        assert_matches!(result.unwrap(), Outcome::NoChanges);
     }
 
     #[test]
-    fn enable_nonexistent_alias() {
+    fn disable_nonexistent_alias() {
         let mut config = sample_config();
-        let result = enable_alias(&mut config, "nonexisting");
+        let result = disable_alias(&mut config, "nonexisting");
         assert!(result.is_err());
         assert_eq!(result.err().unwrap(), Failure::AliasDoesNotExist);
     }
 
     #[test]
-    fn enable_alias_in_disabled_group() {
+    fn disable_alias_in_disabled_group() {
         let mut config = sample_config();
-        let result = enable_alias(&mut config, "alias2");
+        let result = disable_alias(&mut config, "alias2");
         assert!(result.is_ok());
-        assert!(config.aliases["alias2"].enabled);
+        assert!(!config.aliases["alias2"].enabled);
         assert_eq!(result.unwrap(), Outcome::ConfigChanged);
     }
 
     #[test]
-    fn enable_nonexistent_group() {
+    fn disable_disabled_group() {
         let mut config = sample_config();
-        let result = enable_group(&mut config, "nonexisting", &ShellType::Bash);
+        let result = disable_group(&mut config, "disabled_group", &ShellType::Bash);
+        assert!(result.is_ok());
+        assert!(!config.groups["disabled_group"]);
+        assert_eq!(result.unwrap(), Outcome::NoChanges);
+    }
+
+    #[test]
+    fn disable_nonexistent_group() {
+        let mut config = sample_config();
+        let result = disable_group(&mut config, "nonexisting", &ShellType::Bash);
         assert!(result.is_err());
         assert_eq!(result.err().unwrap(), Failure::GroupDoesNotExist);
     }
 
     #[test]
-    fn enable_enabled_group() {
+    fn disable_empty_group() {
         let mut config = sample_config();
-        let result = enable_group(&mut config, "enabled_group", &ShellType::Bash);
+        let result = disable_group(&mut config, "empty_group", &ShellType::Bash);
         assert!(result.is_ok());
-        assert!(config.groups["enabled_group"]);
-        assert_eq!(result.unwrap(), Outcome::NoChanges);
-    }
-
-    #[test]
-    fn enable_empty_group() {
-        let mut config = sample_config();
-        let result = enable_group(&mut config, "empty_group", &ShellType::Bash);
-        assert!(result.is_ok());
-        assert!(config.groups["empty_group"]);
+        assert!(!config.groups["empty_group"]);
         assert_eq!(result.unwrap(), Outcome::ConfigChanged);
     }
 
     #[test]
-    fn enable_group_with_disabled_aliases() {
+    fn disable_group_with_enabled_aliases() {
         let mut config = sample_config();
-        let result = enable_group(&mut config, "disabled_group", &ShellType::Bash);
+        let result = disable_group(&mut config, "enabled_group", &ShellType::Bash);
         assert!(result.is_ok());
-        assert!(config.groups["disabled_group"]);
-        assert_eq!(result.unwrap(), Outcome::ConfigChanged);
-    }
-
-    #[test]
-    fn enable_group_with_enabled_aliases() {
-        let mut config = sample_config();
-        let _ = enable_alias(&mut config, "alias2");
-        assert!(config.aliases["alias2"].enabled);
-
-        let result = enable_group(&mut config, "disabled_group", &ShellType::Bash);
-        assert!(result.is_ok());
-        assert!(config.groups["disabled_group"]);
+        assert!(!config.groups["enabled_group"]);
         assert_matches!(result.unwrap(), Outcome::Command(_));
+    }
+
+    #[test]
+    fn disable_group_with_disabled_aliases() {
+        let mut config = sample_config();
+        let _ = disable_alias(&mut config, "alias1");
+        assert!(!config.aliases["alias1"].enabled);
+
+        let result = disable_group(&mut config, "enabled_group", &ShellType::Bash);
+        assert!(!config.groups["enabled_group"]);
+        assert_eq!(result.unwrap(), Outcome::ConfigChanged);
     }
 }